Rename FatPoints to InstancedPoints
