+++ examples/files.json
@@ -323,6 +323,7 @@
 		"webgpu_depth_texture",
 		"webgpu_equirectangular",
 		"webgpu_instance_mesh",
+		"webgpu_instance_points",
 		"webgpu_instance_uniform",
 		"webgpu_lights_custom",
 		"webgpu_lights_ies_spotlight",
@@ -339,7 +340,6 @@
 		"webgpu_morphtargets",
 		"webgpu_occlusion",
 		"webgpu_particles",
-		"webgpu_points_fat",
 		"webgpu_rtt",
 		"webgpu_sandbox",
 		"webgpu_shadowmap",
