+++ examples/jsm/points/InstancedPointsGeometry.js
@@ -9,15 +9,15 @@ import {
 
 const _vector = new Vector3();
 
-class FatPointsGeometry extends InstancedBufferGeometry {
+class InstancedPointsGeometry extends InstancedBufferGeometry {
 
 	constructor() {
 
 		super();
 
-		this.isFatPointsGeometry = true;
+		this.isInstancedPointsGeometry = true;
 
-		this.type = 'FatPointsGeometry';
+		this.type = 'InstancedPointsGeometry';
 
 		const positions = [ - 1, 1, 0, 1, 1, 0, - 1, - 1, 0, 1, - 1, 0 ];
 		const uvs = [ - 1, 1, 1, 1, - 1, - 1, 1, - 1 ];
@@ -171,4 +171,4 @@ class FatPointsGeometry extends InstancedBufferGeometry {
 
 }
 
-export { FatPointsGeometry };
+export { InstancedPointsGeometry };
