+++ examples/screenshots/webgpu_instance_points.jpg
